Fix for track gaps and static track heights
The navigation panel didn't really respect track gaps. When rendering
the navigation bar, there were some shenanigans going on with the
height and y-position (essentially silently tripling the values when
you ask it to render a track bar). The tripling has been moved
explicitly to the call from the actual renderer, and removed for things
like static track heights, simplifying the calls for rendering a track
bar.


Former-commit-id: f164f4b7f674ea31bf11871a10cee31a74877e80