Fix for reusing the reader object
When you re-use a reader object, and there is no data returned
it used to just return the existing result object with the reader
when the data got spat back. This fixes it so that if you're using
a callback passed with the retrieve function, it will clear out
the result object first, before sending back the data.


Former-commit-id: 37a5b5777be9e01a985a0fe1d951e8bf5f8e4aa3