Fix up the end of touch events for button zooming
When you ended a zoom on a touch device where you had
used the buttons, it wouldn't fire the completion of the
zoom gesture. This fixes up the logic so that if you
do have a "center" zoomCenter value, it remembers that
you haven't set it, and doesn't wait for the gesture
to end.


Former-commit-id: f7dcec8e289c43539e7240834c1111f6453beeb5