Add support for printing on the sequence renderer
While there was some support for printing on the previous
versions of the renderer, this revision allows the
sequence to be automatically re-sized to the page size.
To achieve this, you need to have the grow_container attribute
unset (so the renderer doesn't try and adapt the size of the
canvas to the size of the container).


Former-commit-id: 38f2dc630359085ad6f970ddc7507a85ac728960